assembler: introduce operator 'LAX'
Loading an address greater than 8 bits in length was not possible with
the 'LD' operator since the address operand had only 8 bits left to
fill. Therefore, introduce the operator 'LAX' to load an immediate
address having a maximum length of 12 bits. 'LAX' loads the address
implicitly into a special register known as the accumulator which is not
visible to the end-user. Now, instructions such as:

    LAX #512

are possible.

Signed-off-by: Shourya Shukla <periperidip@gmail.com>