assembler: introduce operator 'STX'
Introduce the operator 'STX' to store a value directly from the implicit
register 'AX' into a specified register. The instruction to be used is:

    STX R2

if the loading is to be done into the register 'R2'. Now, loading a
value greater than 8 bits into a register is a 2-step process due to the
previous commit '645e2f2 (assembler: introduce operator 'LAX',
2022-02-03)' and this commit. To load a value such as '512' into the
register R2:

    LAX #512    ; Use the value as an immediate
    STX R2      ; The value is stored from AX into R2

Reflect this in the 'asm/fibonacci.s' program as well.

Signed-off-by: Shourya Shukla <periperidip@gmail.com>